Complete exercise 4.21

+++ controllers/blogs.js
@@ -60,8 +60,31 @@ blogsRouter.get('/:id', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const token = request.token
+  if (!token) {
+    return response.status(401).json({error: 'token is null'})
+  }
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({error: 'token missing or invalid'})
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(400).json({error: 'requested blog to delete does not exist'})
+  }
+  const user = await User.findById(decodedToken.id)
+  if (!user) {
+    return response.status(400).json({error: 'user could not be found matching token in request'})
+  }
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({error: 'user is not able to delete this blog'})
+  }
+
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
